+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# 0.7.2 
+# 0.7.2 - February 19, 2024
 - Make cloze deletions more consistent (still not perfect)
 - AnkiBrain will now use its own note types, AnkiBrain-Basic and AnkiBrain-Cloze. This fixes the bug for non-English
 users for adding cards to a deck.